Updated empty example project_{device}.js files
The user is told to copy project_{device}.js files when creating a new
PRU project, however "empty" is hardcoded in post build steps -- leading
to conflicts when attempting to build and run. I changes this to make
property.name which the user is instructed to change when making their
new project.

Signed-off-by: baileylooper-ti <b-looper@ti.com>